my project test commit

+++ controller/property.controller.js
@@ -1,13 +1,10 @@
-const {hashPassword,generateNumericOTP} = require("../utils/helperFunc")
+const {hashPassword,generateNumericOTP,isAdmin} = require("../utils/helperFunc")
 const Property = require("../models/property.models")
 const cloudinary = require("../utils/cloudinary")
+const {h } = require("../utils/helperFunc")
 
 const {
-    validateUserRegisterInput, 
-    validateUserLoginInput,
-    validateOtpInput,
-    validateVEmailInput,
-    validateResetPasswordInput,
+ 
     propertyValidationSchema
     } = require('../validator/validator')
     const path = require("path")
@@ -21,18 +18,22 @@ class PropertyController{
     
 async addProperty(req, res) {
 try {
-    // Validate incoming data
+  let  is_admin = await isAdmin()
+  if(!is_admin)return res.status(400).json({ success:false, data:{message: "only admin can perform this operation"}});
     const { error } = propertyValidationSchema.validate(req.body);      
           // If validation fails, return an error response
           if (error) {
             return res.status(400).json({ status: false, data: { message: error.details[0].message } });
           }  
+
     const property = {
         ...req.body,
         user:req.user.id,
         state:"publish",
         investment: req.body.investment
     };
+
+
     let newProp = await Property.create(property)
     // Respond with success
     return res.status(201).json({
@@ -47,6 +48,8 @@ try {
 async getProperties(req, res) {
     try {
       const type = req.query.type
+     let is_admin = await isAdmin()
+      if(!is_admin)return res.status(400).json({ success:false, data:{message: "only admin can perform this operation"}});
       const properties = await Property.find({property_type:type});
       return res.status(200).json({success:true,data:{ properties}});
     } catch (error) {
@@ -57,10 +60,14 @@ async getProperties(req, res) {
   // Get Property by ID
   async getPropertyById(req, res) {
     try {
+      let is_admin = await isAdmin()
+      if(!is_admin)return res.status(400).json({ success:false, data:{message: "only admin can perform this operation"}});
       const property = await Property.findById(req.params.id);
+
       if (!property) {
         return res.status(404).json({ success:false, data:{message: 'Property not found'}});
       }
+
       return res.status(200).json({success:true,data:{property}});
     } catch (error) {
       return res.status(400).json({ success:false, data:{message: error.message}});
@@ -70,6 +77,8 @@ async getProperties(req, res) {
   // Update Property by ID
   async updateProperty(req, res) {
     try {
+      let is_admin = await isAdmin()
+      if(!is_admin)return res.status(400).json({ success:false, data:{message: "only admin can perform this operation"}});
       const { error } = propertyValidationSchema.validate(req.body);      
       // If validation fails, return an error response
       if (error) {
@@ -103,7 +112,10 @@ async getProperties(req, res) {
   // Delete Property by ID
   async deleteProperty(req, res) {
     try {
+      let is_admin = await isAdmin()
+      if(!is_admin)return res.status(400).json({ success:false, data:{message: "only admin can perform this operation"}});
       const deletedProperty = await Property.deleteOne({_id:req.params.id});
+
       if (!deletedProperty) {
         return res.status(404).json({ success:false, data:{message: 'Property not found'}});
       }
